Move loader widget
