+++ src/window/widgets/Loader.svelte
